+++ test/orchard/inspect_test.clj
@@ -896,14 +896,14 @@
                       (datafy-section rendered))))))))
 
 
-(deftest tap-current-value []
+(deftest tap-current-value
   (testing "tap> current value"
-    (let [_ (atom nil)
-          test-tap-handler (fn [x] (reset! _ x))]
+    (let [proof (atom nil)
+          test-tap-handler (fn [x] (reset! proof x))]
 
       (add-tap test-tap-handler)
       (-> (inspect/fresh)
           (inspect/start 123)
           (inspect/tap-current-value))
       (remove-tap test-tap-handler)
-      (is (= 123 @_)))))
+      (is (= 123 @proof)))))